requirements.txt
# Fiduciary Duty Legal Eval - Python Dependencies
# Install with: uv pip install -r requirements.txt

openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
rapidfuzz>=3.0.0
pydantic>=2.0.0
